update: move convert function to src
